+++ app/src/main/java/remix/myplayer/ui/widget/fastcroll_recyclerview/LocationRecyclerView.kt
@@ -47,7 +47,7 @@ class LocationRecyclerView @JvmOverloads constructor(
    *
    * @param data
    */
-  private fun smoothScrollTo(data: List<Song>, currentId: Int) {
+  private fun smoothScrollTo(data: List<Song>, currentId: Long) {
     for (i in data.indices) {
       if (data[i].id == currentId) {
         pos = i
